Restructured NH Group Member Opt Out option
1. Removed all the opt-out options except for "No longer attending"
2. Once a group member is removed from a group a workflow will no longer be set off.
3. A note will be placed onto removed group members profile.
4. Removed all unused blocks
