+++ church.ccv.Utility/church.ccv.Utility.csproj
@@ -65,9 +65,6 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Group\ToolboxGroupDetailLava.ascx.cs">
-      <SubType>ASPXCodeBehind</SubType>
-    </Compile>
     <Compile Include="Group\ToolboxGroupMemberDetail.ascx.cs">
       <SubType>ASPXCodeBehind</SubType>
     </Compile>
